template/{{projectSlug}}/requirements/local.txt
-r base.txt

Werkzeug[watchdog]==3.0.1
psycopg[binary]==3.1.13
{%- if useCelery %}
watchfiles==0.21.0
{%- endif %}

# Testing
# ------------------------------------------------------------------------------
mypy==1.6.1
django-stubs[compatible-mypy]==4.2.6
pytest==7.4.3
pytest-sugar==0.9.7
{%- if restFramework == 'DRF' %}
djangorestframework-stubs[compatible-mypy]==3.14.4
{%- endif %}

# Code quality
# ------------------------------------------------------------------------------
flake8==6.1.0
flake8-isort==6.1.1
coverage==7.3.2
black==23.11.0
djlint==1.34.0
pylint-django==2.5.5
{%- if useCelery %}
pylint-celery==0.3
{%- endif %}
pre-commit==3.5.0

# Django
# ------------------------------------------------------------------------------
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-coverage-plugin==3.1.0
pytest-django==4.7.0
